// web/types/server/VpnDevice.ts
// This file was generated by [ts-rs](https://github.com/Aleph-Alpha/ts-rs). Do not edit this file manually.
import type { VpnDeviceType } from "./VpnDeviceType";

export type VpnDevice = { type: VpnDeviceType, name: string, ip_addresses: Array<string>, last_seen: string | null, created_at: string | null, online: boolean, };